feat: the user can pass custom arguments to his TorchAlgo instance (#130)
