+++ connectlib/algorithms/pytorch/torch_scaffold_algo.py
@@ -52,6 +52,10 @@ class TorchScaffoldAlgo(TorchAlgo):
     :py:func:`~connectlib.algorithms.pytorch.torch_scaffold_algo.TorchScaffoldAlgo._local_predict` methods,
     and can override other methods if necessary.
 
+    To add a custom parameter to the ``__init__``of the class, also add it to the call to ``super().__init__```
+    as shown in the example with ``my_custom_extra_parameter``. Only primitive types (str, int, ...) are supported
+    for extra parameters.
+
     Example:
 
         .. code-block:: python
@@ -59,6 +63,7 @@ class TorchScaffoldAlgo(TorchAlgo):
             class MyAlgo(TorchScaffoldAlgo):
                 def __init__(
                     self,
+                    my_custom_extra_parameter,
                 ):
                     super().__init__(
                         model=perceptron,
@@ -68,7 +73,8 @@ class TorchScaffoldAlgo(TorchAlgo):
                         index_generator=NpIndexGenerator(
                             num_updates=10,
                             batch_size=32,
-                        )
+                        ),
+                        my_custom_extra_parameter=my_custom_extra_parameter,
                     )
                 def _local_train(
                     self,
@@ -125,6 +131,8 @@ class TorchScaffoldAlgo(TorchAlgo):
         scheduler: Optional[torch.optim.lr_scheduler._LRScheduler] = None,
         with_batch_norm_parameters: bool = False,
         c_update_rule: CUpdateRule = CUpdateRule.FAST,
+        *args,
+        **kwargs,
     ):
         """The ``__init__`` function is called at each call of the ``train`` or ``predict`` function
         For round>2, some attributes will then be overwritten by their previous states in the `load()` function,
@@ -156,6 +164,8 @@ class TorchScaffoldAlgo(TorchAlgo):
             optimizer=optimizer,
             index_generator=index_generator,
             scheduler=scheduler,
+            *args,
+            **kwargs,
         )
         if self._index_generator.num_updates <= 0:
             raise NumUpdatesValueError("Num_updates needs to be superior to 0 for TorchScaffoldAlgo.")
