+++ connectlib/remote/register/__init__.py
@@ -1,0 +1,5 @@
+from connectlib.remote.register.register import register_algo
+
+__all__ = [
+    "register_algo",
+]
