+++ docs/conf.py
@@ -188,6 +188,8 @@ nitpick_ignore = [
     ("py:class", "substra.sdk.client.Client"),
     ("py:class", "substra.sdk.models.ComputePlan"),
     ("py:class", "ComputePlan"),
+    ("py:class", "substratools.algo.CompositeAlgo"),
+    ("py:class", "substratools.algo.AggregateAlgo"),
 ]
 
 html_css_files = [
